Updated to compile with Rust 1.0.0-alpha

+++ src/data.rs
@@ -1,5 +1,5 @@
 // Copyright (c) 2014-2015 by Michael Dvorkin. All Rights Reserved.
 
 // Useful square constants.
-pub const A1: uint = 0;
-pub const H8: uint = 63;
+pub const A1: usize = 0;
+pub const H8: usize = 63;
